FileExt.SafeDelete no error when file does not exist

+++ Dinah.Core/_IO/FileExt.cs
@@ -6,13 +6,18 @@ namespace Dinah.Core.IO
 {
     public static class FileExt
     {
+        /// <summary>Delete file. No error when file does not exist.
+        /// Exceptions are logged, not thrown.</summary>
+        /// <param name="source">File to delete</param>
         public static void SafeDelete(string source)
         {
+            if (!File.Exists(source))
+                return;
+
             while (true)
             {
                 try
                 {
-                    // deletes file if it exists. no error if it doesn't exist
                     File.Delete(source);
 					Serilog.Log.Logger.Information($"File successfully deleted: {source}");
                     break;
@@ -25,7 +30,13 @@ namespace Dinah.Core.IO
             }
         }
 
-        public static void SafeMove(string source, string target)
+        /// <summary>
+		/// Moves a specified file to a new location, providing the option to specify a newfile name.
+		/// Exceptions are logged, not thrown.
+		/// </summary>
+		/// <param name="source">The name of the file to move. Can include a relative or absolute path.</param>
+		/// <param name="target">The new path and name for the file.</param>
+		public static void SafeMove(string source, string target)
         {
             while (true)
             {
